footer/nav data refactor
