+++ src/components/Menu.js
@@ -18,13 +18,12 @@ class MenuPages extends Component {
                         {this.props.menuData.map((item, index) => (
                             <div key={index}
                                 className="col-xl-4 col-md-6 pt-lg-0 product-content-wrapper-lg container-fluid-fluid py-3 pt-5 menu-item-col">
-                                {/* State can be assigned on link, keep here if needed <Link className="d-block" to={item.linkTo} state={ item.name.replace(/\s/g, '') }> */}
                                 <Link className="d-block" to={item.linkTo}>
                                     <div className="primary-subheading-wrapper">
                                         <div className="h2">{item.name}</div>
                                     </div>
                                     <div>
-                                        {/* @TODO Fix the largeplates image bug here, ALSO, remove category image (pointless) and restructure deskop/mobile images folders. */}
+                                        {/* @TODO remove category image (pointless) and restructure deskop/mobile images folders. */}
                                         <img loading="lazy" className="w-100 subheading-img"
                                             src={`/assets/foodImages/categoryImages/desktop/500x375${item.img}`}
                                             alt="" />
