--- src/data/index.js
@@ -1,6 +1,0 @@
-export { footerItemsSecondary } from './footerItemsSecondary';
-export { footerItemsDataMain } from './footerItemsDataMain';
-export { franchiseNavItems } from './franchiseNavItems';
-export { socialNavItems } from './socialNavItems';
-export { navItems } from './navItems';
-
