+++ src/data/linksData/index.js
@@ -1,0 +1,5 @@
+export { footerItemsSecondary } from './footerItemsSecondary';
+export { footerItemsDataMain } from './footerItemsDataMain';
+export { franchiseNavItems } from './franchiseNavItems';
+export { socialNavItems } from './socialNavItems';
+export { navItems } from './navItems';
